+++ HISTORY.rst
@@ -5,6 +5,14 @@ Release History
 ---------------
 
 
+Release 0.8.7 (2014-09-10)
+++++++++++++++++++
++ Update README
++ When ``New Project``, no need to select project
++ Fix bug ``c:`` completion
++ Next release: Deploy chosen code file
+
+
 Release 0.8.6 (2014-09-09)
 ++++++++++++++++++
 + Add ``c:`` prefix for custom component completion